#7470 Fixes a large amount of bugs and incorrectly refactored functions
This update fixes many issues that exist within the transformation layer and other functions. It also
refactors the transformation layer to be more concise and compartmentalized. Several issues with
the CLI are resolved and dependencies cleaned up while adding some more documentation.
It also ensures the Hough transform is as close as possible to the original while fixing any of the
problems encountered with the OpenCV library.

* Fixes a the majority of transformations resolved in the LUT creation.
* Fixes the Hough transform to be as close as possible while based on OpenCV.
* Fixes the OpenCV ML library usage.
* Improves documentation.
* Fixes CLI errors.

+++ SlideStandardization/MiscFunctionality.cpp
@@ -5,10 +5,10 @@
 
 namespace ASAP::MiscFunctionality
 {
-	std::vector<uint32_t> CreateListOfRandomIntegers(size_t size)
+	std::vector<size_t> CreateListOfRandomIntegers(size_t size)
 	{
-		std::vector<uint32_t> random_numbers(size);
-		for (uint32_t element = 0; element < size; ++element)
+		std::vector<size_t> random_numbers(size);
+		for (size_t element = 0; element < size; ++element)
 		{
 			random_numbers[element] = element;
 		}
